Added a way to convert continuous features to discrete ones.
Removed ValueMap and ReverseValueMap from the Feature definition.
Added CreateDiscrete function to Feature to make up for their removal.
Added a StringValue to the feature Instances.
Added a way to add a column to the Table class.
Now converts all continuous data into discrete data before passing into
the model.
Removed processing of Continious Data in the discision tree as the
discrete versions performed better.
Turned off the SampleWeight feature as it was basically noise.